Add features & fix bugs, so so much

+++ Ledger/Utilities/YGTools.h
@@ -48,7 +48,7 @@
 
 + (CGFloat)defaultFontSize;
 
-+ (NSInteger)deviceScreenWidth;
++ (CGFloat)deviceScreenWidth;
 
 + (UIColor *)colorForActionDisable;
 + (UIColor *)colorForActionDeactivate;
